+++ NEWS.md
@@ -3,7 +3,11 @@
 * Extensive linting. Passes expect_lint_free
 * Removed support for caption tracks from old Youtube API as client should only be for V3. Changed the get_captions API.
 * support the deletes
-* translate_filter in get_comment_threads also supports pagetoken, which it didn't. 
+* translate_filter in get_comment_threads also supports pagetoken, which it didn't.
+* removes cats (prints) from get_stats based on user feedback
+* list_channel_videos now supports getting all the videos from the playlists
+* get_playlist_items supports simplify, defaults to simplify, and also allows getting all the videos from the playlist easily.
+
 
 # version 0.8.0